test: setting up basic structure for test comparing generated spin gif with reference datauri
