+++ src/components/EmojiPanel/index.js
@@ -56,6 +56,7 @@ const EmojiPanel = ({
       <Image
         src={gif}
         loading={loading}
+        data-emoji={name}
         alt={`The generated ${name} animated emoji`}
       />
       <Name>:{name}:</Name>
