+++ test/spin-gif-datauri.js
